style: reformat code with scalafmt
