+++ src/main/scala/dev/mongocamp/driver/mongodb/operation/CrudObserver.scala
@@ -30,10 +30,4 @@ trait CrudObserver[A] extends Crud[A] {
 
 }
 
-class SimpleObserver[T] extends Observer[T] with LazyLogging {
-  override def onError(e: Throwable): Unit = logger.error(e.getMessage, e)
 
-  override def onComplete(): Unit = {}
-
-  override def onNext(result: T): Unit = {}
-}
