+++ src/main/scala/dev/mongocamp/driver/mongodb/relation/Relationship.scala
@@ -11,25 +11,3 @@ abstract class Relationship {
     removeCachedValue(key)
   }
 }
-
-case class OneToOneRelationship[A](dao: MongoDAO[A], daoKey: String, useCache: Boolean = true) extends Relationship {
-
-  def relatedRecord(value: Any): Option[A] = {
-    val key = "%s_%s".format(id, value)
-    if (!useCache || !hasCachedValue(key))
-      addCachedValue(key, dao.find(daoKey, value).resultOption())
-    getCachedValue[Option[A]](key)
-  }
-
-}
-
-case class OneToManyRelationship[A](dao: MongoDAO[A], daoKey: String, useCache: Boolean = true) extends Relationship {
-
-  def relatedRecords(value: Any): List[A] = {
-    val key = "%s_%s".format(id, value)
-    if (!useCache || !hasCachedValue(key))
-      addCachedValue(key, dao.find(daoKey, value).resultList())
-    getCachedValue[List[A]](key)
-  }
-
-}
